Application Scaffolding: Refactored the project directories
- Created packages for student, tutor, and app fragments.
- Moved corresponding files into the packages.
