+++ app/src/main/java/cs/dal/krush/TutorMainActivity.java
@@ -7,6 +7,10 @@ import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentTransaction;
 import android.view.MenuItem;
 import cs.dal.krush.helpers.BottomNavigationViewHelper;
+import cs.dal.krush.tutorFragments.TutorCalendarFragment;
+import cs.dal.krush.tutorFragments.TutorHomeFragment;
+import cs.dal.krush.tutorFragments.TutorProfileFragment;
+import cs.dal.krush.tutorFragments.TutorSessionsFragment;
 
 
 public class TutorMainActivity extends FragmentActivity {
