+++ app/src/main/java/cs/dal/krush/appFragments/SessionItemFragment.java
@@ -1,4 +1,5 @@
-package cs.dal.krush;
+package cs.dal.krush.appFragments;
+
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -6,11 +7,19 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class StudentHomeFragment extends Fragment {
+import cs.dal.krush.R;
+
+
+/**
+ * This view is a generic view for displaying one session for the Session History list.
+ */
+public class SessionItemFragment extends Fragment {
+
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.student_home, container, false);
+        return inflater.inflate(R.layout.session_item, container, false);
     }
+
 }
