+++ app/src/main/java/cs/dal/krush/studentFragments/StudentHomeFragment.java
@@ -1,4 +1,4 @@
-package cs.dal.krush;
+package cs.dal.krush.studentFragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -6,16 +6,13 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-/**
- * Student quick book view.
- */
-public class StudentQuickBookFragment extends Fragment {
+import cs.dal.krush.R;
 
+public class StudentHomeFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.student_quickbook, container, false);
+        return inflater.inflate(R.layout.student_home, container, false);
     }
-
 }
