+++ app/src/main/java/cs/dal/krush/studentFragments/StudentProfileFragment.java
@@ -1,4 +1,4 @@
-package cs.dal.krush;
+package cs.dal.krush.studentFragments;
 
 
 import android.os.Bundle;
@@ -6,6 +6,8 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import cs.dal.krush.R;
 
 
 /**
