+++ app/src/main/java/cs/dal/krush/studentFragments/StudentQuickBookFragment.java
@@ -1,4 +1,4 @@
-package cs.dal.krush;
+package cs.dal.krush.studentFragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -6,11 +6,18 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class TutorProfileFragment extends Fragment {
+import cs.dal.krush.R;
+
+/**
+ * Student quick book view.
+ */
+public class StudentQuickBookFragment extends Fragment {
+
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.tutor_profile, container, false);
+        return inflater.inflate(R.layout.student_quickbook, container, false);
     }
+
 }
