+++ app/src/main/java/cs/dal/krush/studentFragments/StudentSessionsFragment.java
@@ -1,10 +1,12 @@
-package cs.dal.krush;
+package cs.dal.krush.studentFragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import cs.dal.krush.R;
 
 /**
  * Student Sessions History view.
