+++ app/src/main/java/cs/dal/krush/tutorFragments/TutorCalendarFragment.java
@@ -1,23 +1,23 @@
-package cs.dal.krush;
+package cs.dal.krush.tutorFragments;
 
-
-import android.os.Bundle;
 import android.support.v4.app.Fragment;
+import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import cs.dal.krush.R;
 
 /**
- * This view is a generic view for displaying one session for the Session History list.
+ * Tutor calendar view.
  */
-public class SessionItemFragment extends Fragment {
+public class TutorCalendarFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.session_item, container, false);
+        return inflater.inflate(R.layout.tutor_calendar_fragment, container, false);
     }
 
 }
