+++ app/src/main/java/cs/dal/krush/tutorFragments/TutorProfileFragment.java
@@ -1,5 +1,4 @@
-package cs.dal.krush;
-
+package cs.dal.krush.tutorFragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -7,17 +6,13 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import cs.dal.krush.R;
 
-/**
- * A simple {@link Fragment} subclass.
- */
-public class StudentQuickBookItemFragment extends Fragment {
-
+public class TutorProfileFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.student_quick_book_item, container, false);
+        return inflater.inflate(R.layout.tutor_profile, container, false);
     }
-
 }
