+++ app/src/main/java/cs/dal/krush/tutorFragments/TutorSessionsFragment.java
@@ -1,21 +1,22 @@
-package cs.dal.krush;
+package cs.dal.krush.tutorFragments;
 
-import android.support.v4.app.Fragment;
 import android.os.Bundle;
+import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import cs.dal.krush.R;
+
 /**
- * Tutor calendar view.
+ * Tutor Sessions History view.
  */
-public class TutorCalendarFragment extends Fragment {
+public class TutorSessionsFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.tutor_calendar_fragment, container, false);
+        return inflater.inflate(R.layout.tutor_sessions, container, false);
     }
-
 }
