+++ app/src/main/res/layout/session_item.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cs.dal.krush.SessionItemFragment">
+    tools:context="cs.dal.krush.appFragments.SessionItemFragment">
 
     <TextView
         android:id="@+id/sessionItemTitle"
