+++ app/src/main/res/layout/student_profile.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cs.dal.krush.StudentProfileFragment"
+    tools:context="cs.dal.krush.studentFragments.StudentProfileFragment"
     android:id="@+id/student_profile">
 
     <TextView
