+++ app/src/main/res/layout/student_quick_book_item.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cs.dal.krush.StudentQuickBookItemFragment">
+    tools:context="cs.dal.krush.studentFragments.StudentQuickBookItemFragment">
 
     <TextView
         android:id="@+id/quickBookItemTitle"
