+++ app/src/main/res/layout/student_quickbook.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cs.dal.krush.StudentQuickBookFragment"
+    tools:context="cs.dal.krush.studentFragments.StudentQuickBookFragment"
     android:id="@+id/student_quickbook">
 
     <TextView
