+++ app/src/main/res/layout/student_sessions.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cs.dal.krush.StudentSessionsFragment"
+    tools:context="cs.dal.krush.studentFragments.StudentSessionsFragment"
     android:id="@+id/student_sessions">
 
     <TextView
