+++ app/src/main/res/layout/tutor_calendar_fragment.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="cs.dal.krush.TutorCalendarFragment">
+    tools:context="cs.dal.krush.tutorFragments.TutorCalendarFragment">
 
     <com.prolificinteractive.materialcalendarview.MaterialCalendarView
         xmlns:app="http://schemas.android.com/apk/res-auto"
