+++ app/src/main/res/layout/tutor_profile.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cs.dal.krush.TutorProfileFragment"
+    tools:context="cs.dal.krush.tutorFragments.TutorProfileFragment"
     android:id="@+id/tutor_profile">
 
     <TextView
