+++ app/src/main/res/layout/tutor_sessions.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cs.dal.krush.TutorSessionsFragment"
+    tools:context="cs.dal.krush.tutorFragments.TutorSessionsFragment"
     android:id="@+id/tutor_sessions">
 
     <TextView
